Scheduler: match OS and ISA between job and worker
With these changes the `platform` attributes of the Commands and
workers start to be considered by the scheduler.

When a given worker requests a Job, the scheduler will now try and assign it
the first job in the queue that can be executed in that worker.

As a side effect, submitting an Action with no Command now will result
in a FailedPreconditionError exception.

(This will allow to address #104)