CAS.{Up, Down}loader(): if provided, respect `CacheCapabilities`
Try fetching the `CacheCapabilities.max_batch_total_size_bytes` value
from the server and, if the server provides a limit, respect it for
batch transactions.

Also, make the threshold value for determining when to queue a file
for fetching in a batch request a configurable % of the maximum
batch request size.

(Addresses #100)