Server: add support for RequestMetadata
The `Scheduler` now keeps track of that information and could employ
them for logging and monitoring purposes when a remote session is
finished.

This involved some refactoring and the introduction of a new
`Peer` class, which also keeps track of all the `Peer` instances
(indexed by `Peer.uid` values).

(Addresses #106)