tenantmanager.py: Handle exceptions in complete_lease
The complete_lease callback currently just assumes that the task was
completed successfully. In situations where an exception was raised,
`task.result()` re-raises the exception and so the lease state never
gets updated. This makes the job hang forever when there is a
bot-related error.

This commit adds logic to check whether the task failed, and if so set
the lease status and state appropriately. There is also a small change
to the server-side result logic to allow empty results to be accepted,
since sometimes the exception can occur before there is useful content
for an ActionResult.