Fix some docs warnings
There are a bunch of warnings in the docs build, some more important
than others. One of these (the conf.py fix) was breaking the docs
build entirely for sphinx 2.0.

This fixes most of the warnings that aren't in the Python files
generated from the .proto files. The only unfixed warning that
isn't from generated code is about the FAQ not being in a table
of contents anywhere, but that seems to be a deliberate choice and
is harmless to ignore.