Add a tox.ini file to ease management of dependencies
Currently, running BuildGrid locally for development or testing purposes
means either manually using virtualenvs, or rebuilding a docker
container every time a change is made.

This adds a tox.ini file which defines targets for running the tests,
building the docs, running a basic server and bot, and a generic "venv"
target for running arbitrary commands in an environment with BuildGrid
and its dependencies installed.

This allows faster turnaround time when testing changes to the code by
avoiding docker rebuilds, without the overhead of manually working with
virtualenvs or installing things globally.