Don't attempt to update read-only caches
If the ActionCache is in read-only mode, the scheduler will still try to
cache successful job results (unless the peer advised otherwise). This
will cause an endless loop of uncaught NotImplementedErrors which just
seems like the build has hung from the client-side, since the operation
stage never gets updated.

This commit checks whether updating the cache is allowed before
attempting to do so, avoiding raising the NotImplementedError when using
a read-only cache.