Implement timestamps for the buildbox bot
This makes use of the --output-times command line argument to buildbox
to make it output a file containing an ExecutedActionMetadata message.

This file is then read by the bot and the contents added to the
ActionResult, making the timestamps provided by the buildbox bot
consistent with the other provided bots.

This solves part of #83.