scheduler.py: allow requests to block until a job appears
This replaces the list used as a queue in the scheduler with a Queue,
which supports waiting for objects to be added to it. This allows
request_job_leases to block waiting for relevant jobs to appear until a
specified timeout.

This is mostly based on d0a84116 from the earlier smart bot calls MR[0].

[0] https://gitlab.com/BuildGrid/buildgrid/merge_requests/113