Update the bot to utilise blocking calls to request_leases
The scheduler supports blocking until either a job is available or a
timeout is reached when requesting leases, rather than just checking
once and returning. This commit updates the bot code to take advantage
of this feature, and in doing so switches from rapidly polling the
server to check for jobs to a long-polling implementation with a timeout
which defaults to 30s.

This should reduce the server load when many bots are connected, but
also has a detrimental effect on thread consumption, since each bot
connection requires its own thread.

This commit is heavily based on 0d241b27 from the older smart bot calls
MR (!113).