Add a separate configurable update period during execution
When a bot is executing a job, it should send UpdateBotSession requests
more frequently in order to minimise the amount of time working on
cancelled jobs.

This commit adds a configurable parameter which is used to determine the
time between requests whilst executing a job. It would be even more
efficient to block on a queue waiting for relevant cancellation events,
similarly to the new lease implementation, but that requires some
further refactoring.