Reject new connections if the thread pool is full
Currently it is possible to connect more bots to the grid than there are
available threads for the grpc server. Since this number is configurable
now, this should be a rare occurrence, but it also shouldn't appear like
everything is working fine if attempted.

The default behaviour of grpc is to queue the requests and only execute
those which there is space for. This works for the bots but causes the
grid to get into a state where builds fail due to a lack of resources.

Setting maximum_concurrent_rpcs to equal the size of the thread pool
causes all connections that don't fit to be rejected, which clearly
indicates that the grid is in poor health and also speeds up the time to
an error for peers (since they fail when attempting to connect, rather
than failing during the build).