Allow passing state to the Job constructor
This will allow Jobs to be instantiated with pre-existing state, rather
than always creating an entirely new job. Also, only pass whether or
not to cache the job rather than passing the entire Action, since the
rest isn't needed and is also complicated to obtain if recreating the
Job from an external place.