Pre-emptively set worker_start_time before lease is accepted
Waiting for the lease to be accepted by the worker leads to inaccurate
timestamps being used in the operation metadata. In the case where the
execution service restarts between sending the job to the worker and
receiving the first communication back, this can lead to confusing
results where the reported total execution time is significantly less
than the sum of the parts measured by the worker.

This commit solves this problem by assuming that the worker will accept
the lease and setting the timestamp when sending the lease to the
worker, rather than waiting for acceptance.