Remove association_proxy for dict-like access to platform_requirements
It turns out that the platform requirements are actually a dictionary
with `set` values, to allow for the same key to have many values, so
this proxy doesn't make much sense.

This commit replaces it with a regular relationship and some additional
code to convert between the list-based relationship and the dict-based
internal representation.