Track peers in the scheduler
This commit moves the mapping of peers (and their message queues) to
operations out of the job class and into the scheduler. This means
that the scheduler isn't entirely stateless, but its state that is
only relevant whilst the scheduler is running.

Having the peers stored in the job class doesn't work with the SQL
implementation of DataStoreInterface with a stateless scheduler, since
we don't store the peers and therefore cannot recreate them when
(for example) updating a job.