Store the job ExecuteResponse in CAS
Currently the SQL data store implementation loses the ExecuteResponse
of a completed job between the job being completed and the peers being
informed of the completion. This commit stores that response in the CAS
when a job finishes, and also records the digest in the database, to
allow retrieval of the response when messaging the peers.

The in-memory data store implementation doesn't do anything, but gets a
new parameter to keep an equivalent API. This parameter is silently
ignored. There's no need to write to the CAS with the in-memory data
store, since if the job is lost for whatever reason it is gone forever
anyway.