Add a way to persist the job result
This adds a way to pass an ExecuteResponse to the Job constructor, and
also adds support for storing a message digest in the database for a
job in the SQL data store implementation. This will be used to
reconstruct an ExecuteResponse that was stored in CAS from a scheduler
process that may or may not have seen the response before.