Properly persist operation cancellation
Currently cancellation is persisted at the job level. This is broadly
fine, but causes problems when using a stateless scheduler, since
each scheduler process could have a different opinion on whether a job
is cancelled.

This commit implements persisting the cancelled state of individual
operations, so that the database provides an accurate single source of
truth for whether a job is cancelled or not.