Only include incomplete operations in ListOperations
Currently in the SQL data store implementation, all operations
in the database are returned by ListOperations, which quickly
leads to excessively large responses and slow response times.

This commit extends the query to filter out completed operations,
which matches the behaviour of the in-memory scheduler and improves
the usability of ListOperations with an SQL backend.

The test for operation cancellation is also modified as it relied
on winning a race between the deletion of the operation and the
ListOperations request it makes being handled. The SQL implementation
doesn't leave enough time for this to work, so the test now uses
GetOperation to avoid this race.