Replace use of a PriorityQueue with a list
Since it is impossible to iterate over the contents of a PriorityQueue,
it is unsuitable for the case where the next job in the queue isn't
runnable by a given worker. This will cause the worker to repeatedly
attempt to run the same job, and if no workers can run that job then
no work will get done (barring anything that manages to get assigned
whilst one worker has temporarily removed the problematic job).

Using a list is complicated since we may be adding and removing items
from multiple threads at the same time. Simply locking the list every
time we want to read it is insufficient as it will cause too much delay
with workers long-polling for work.

This commit replaces the PriorityQueue with a list with locks only
obtained when actually modifying the contents of the list. Once the
lock has been obtained but before any modification is done, the list
is checked to make sure that it hasn't changed in the intervening
time. This should ensure that the correct elements are removed when
assigning work.