Refactor get_next_runnable_job to assign leases whilst locked
Currently we get a FOR UPDATE lock in get_next_runnable_job,
but don't actually make any updates whilst holding it. This means
that other workers can get the same job, before it has been marked
as started, so work gets duplicated.

This commit refactors the scheduler to pass a callback to assign
the leases whilst in the FOR UPDATE lock, as the new DataStoreInterface
requires.