Store the worker name for the lease when using the SQL data store
Now that leases can be persisted even after they're completed, it is
useful to see which worker executed which lease. The data model already
supports this, but we never actually update the relevant field.

This commit correctly sets the worker_name field of a lease when
creating it to assign a job to a worker.