Improve test coverage of scheduler.py
At the moment, some key parts of the scheduler are completely untested.
This commit adds some unit tests which provide at least basic coverage
over most of the code, which should make it easier to catch issues and
prevent regressions in the future.

All of the tests added are testing behaviour where regressions were
caused and/or problems encountered in the recent scheduler refactoring.