Store the number of times a job has been attempted in the database
Currently the number of times a job has been retried is lost when using
an SQLDataStore, which means that jobs will be retried forever rather
than respecting the limit set in scheduler.py.

This commit fixes that by adding a field for the number of attempts in
the jobs table, and updating it when undergoing a relevant operation stage
transition.