Switch to using the default NullPool with SQLite
Using a `StaticPool` causes issues with concurrency when
`check_same_thread` is set to False. It does no locking, so allows
multiple threads to concurrently use the connection, which under load
leads to collisions in transactions.

This commit removes that configuration and uses the default pool class,
which is `NullPool` for SQLite. This produces a new connection
whenever one is requested, so the connection is never shared between
threads.

However, SQLite's write performance can mean that queries time out
waiting to acquire a lock when BuildGrid is under load. In order to
alleviate this this commit adds a callback for the 'connect' event when
using SQLite, which sets the journal_mode pragma to `WAL` and the
synchronous pragma to `NORMAL`, which improves write performance at the
expense of some durability. See [0] and [1] for further details.

[0]: https://www.sqlite.org/wal.html
[1]: https://www.sqlite.org/pragma.html#pragma_synchronous