Call notify on our Condition subclass when exiting
This subclass is meant to allow single workers to wait if necessary, but
sometimes they can get stuck waiting if something other than a worker
acquires the underlying lock and doesn't call `notify`.

This commit calls `notify` when calling `__exit__`, so that all uses of
the class as a contextmanager will wake a waiting worker thread (if
any).